refactor(hooks): extract useDebounce, useMediaQuery, useLiquidTextAnimation
Part of #1471. Completes the hooks-layer acceptance criteria.

useDebounce: DocsSearchBar's search-input debounce was the only real
consumer found — NavigationProgress's staggered setTimeout steps
matched an earlier grep for 'Timeout' but aren't a debounce pattern,
left untouched.

useMediaQuery: applied to EscrowFlowDiagram's mobile breakpoint check.
HeroSection's matchMedia usage is absorbed into useLiquidTextAnimation
instead (see below). BackToTopButton's matchMedia call is a one-off
synchronous read in a click handler, not a subscribed value, so it's
left as-is. ThemeProvider's matchMedia usage is left untouched too —
its cleanup calls addEventListener again instead of removeEventListener,
a pre-existing bug that shouldn't be silently fixed inside this refactor.

useLiquidTextAnimation: extracted from HeroSection, its only consumer.
OrchestratorShowcase/WhyStellarSection matched an earlier grep for
'liquid' but only via the word 'liquidity' in copy, not the animation
effect — no actual duplication there.
