+++ src/hooks/useMediaQuery.ts
@@ -1,0 +1,26 @@
+"use client";
+
+import { useEffect, useState } from "react";
+
+/**
+ * Subscribes to a CSS media query and returns whether it currently
+ * matches, updating on change. Matches the pattern previously duplicated
+ * inline via `window.matchMedia(...).matches` + a manual `change`
+ * listener.
+ */
+export function useMediaQuery(query: string): boolean {
+  const [matches, setMatches] = useState(() =>
+    typeof window === "undefined" ? false : window.matchMedia(query).matches
+  );
+
+  useEffect(() => {
+    const mql = window.matchMedia(query);
+    setMatches(mql.matches);
+
+    const handleChange = (e: MediaQueryListEvent) => setMatches(e.matches);
+    mql.addEventListener("change", handleChange);
+    return () => mql.removeEventListener("change", handleChange);
+  }, [query]);
+
+  return matches;
+}
